支持背景高亮，代码块数学公式
